requirements.txt
# Azure Video Transcription Application Requirements
# Install with: pip install -r requirements.txt

# Azure Speech Services SDK
azure-cognitiveservices-speech>=1.23.0
azure-storage-blob>=12.20.0

# Video processing
moviepy==1.0.3

# Environment variables management
python-dotenv>=0.19.0

# Additional dependencies that may be needed
requests>=2.28.0
pathlib>=1.0.1

# Optional: For enhanced audio processing
pydub>=0.25.1

# Optional: For progress bars
tqdm>=4.64.0
